+++ resources/views/markdown/changelog.md
@@ -3,6 +3,12 @@
 Last Update: March 17, 2024\
 Travis Michael Cross <a href="mailto:travis@not.tv">travis@not.tv</a>
 
+## v0.8.7.21
+March 17, 2024
+* Add a job to check a newly uploaded image against the existing image hashes. If it's a duplicate it will update the associated model and queue the image for deletion. Currently the duplicate images are queued for deletion every hour.
+* Update generic image uploader for Admin to handle duplicate images and test the image hash, deletion process.
+
+
 ## v0.8.7.20
 March 17, 2024
 * Improve Hash image, remove duplicate job logic to avoid race conditions.